discription
Signals provide a method of inter process communication in Unix. When a
process receives a signal, its flow of execution is interrupted by the operating
system to call a signal handler. Signals are identified by a number, and each
one has a default signal handler. For example, when CTRL -C is typed in a
program’s controlling terminal, an interrupt signal is sent, which has a default
signal handler that exits the program. This allows the program to be inter-
erupted, even if it is stuck in an infinite loop.


program explainnation
When this program is compiled and executed, signal handlers are
registered, and the program enters an infinite loop. Even though the program
is stuck looping, incoming signals will interrupt execution and call the
registered signal handlers. In the output below, signals that can be triggered
from the controlling terminal are used.

The signal_handler() function,when finished,returns execution back into the interrupted 
loop, whereas the sigint_handler() function exits the program.